requirements.txt
fastapi
uvicorn[standard]
Django
better-profanity
django-cors-headers
pandas
numpy
psycopg2-binary
dj-database-url